// SourceCode/Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/customer/models/customer-care.model.ts
export class CustomerCareModel {
  CustomerCareId: string;
  CustomerCareCode: string;
  EmployeeCharge: string;
  EffecttiveFromDate: Date;
  EffecttiveToDate: Date;
  CustomerCareTitle: string;
  CustomerCareContent: string;
  CustomerCareContactType: string;
  CustomerCareContentSMS: string;
  ExpectedAmount: number;
  IsSendNow: boolean;
  IsEvent: boolean;
  SendDate: Date;
  SendHour: Date;
  CustomerCareEvent: string;
  CustomerCareEventHour: Date;
  CustomerCareContentEmail: string;
  IsSendEmailNow: boolean;
  SendEmailDate: Date;
  SendEmailHour: Date;
  CustomerCareVoucher: string;
  DiscountAmount: number;
  PercentDiscountAmount: number;
  GiftCustomerType1: number;
  GiftTypeId1: string;
  GiftTotal1: number;
  GiftCustomerType2: number;
  GiftTypeId2: string;
  GiftTotal2: number;
  CustomerCareType: number;
  StatusId: string;
  StatusCode: string;
  CreateDate: Date;
  CreateById: string;
  UpdateDate: Date;
  UpdateById: string;
  CustomerCareContactTypeName: string;
  StatusName: string;
  EmployeeChargeName: string;
  IsEditStatus: boolean;
  TypeCustomer: number;

  constructor() {
    this.CustomerCareId = '00000000-0000-0000-0000-000000000000';
    this.CustomerCareCode = '';
    this.EmployeeCharge = null;
    this.EffecttiveFromDate = new Date();
    this.EffecttiveToDate = new Date(d.setDate(d.getDate() + 7));
    this.CustomerCareTitle = null;
    this.CustomerCareContent = null;
    this.CustomerCareContactType = null;
    this.CustomerCareContentSMS = "";
    this.ExpectedAmount = 0;
    this.IsSendNow = null;
    this.IsEvent = null;
    this.SendDate = null;
    this.SendHour = null;
    this.CustomerCareEvent = null;
    this.CustomerCareEventHour = null;
    this.CustomerCareContentEmail = "";
    this.IsSendEmailNow = true;
    this.SendEmailDate = null;
    this.SendEmailHour = null;
    this.CustomerCareVoucher = null;
    this.DiscountAmount = null;
    this.PercentDiscountAmount = null;
    this.GiftCustomerType1 = null;
    this.GiftTypeId1 = null;
    this.GiftTotal1 = null;
    this.GiftCustomerType2 = null;
    this.GiftTypeId2 = null;
    this.GiftTotal2 = null;
    this.CustomerCareType = null;
    this.StatusId = null;
    this.CreateDate = new Date();
    this.CreateById = null;
    this.UpdateDate = new Date();
    this.UpdateById = null;
    this.CustomerCareContactTypeName = null;
    this.StatusName = null;
    this.EmployeeChargeName = null;
    this.IsEditStatus = null;
    this.StatusCode = null;
    this.TypeCustomer = 1;
  }
}